Raw data underlying figures in manuscript
These are all the raw data files underlying the data visualizations presented in the multi-omics characterization of a nitrifying consortium manuscript. All figures the data are keyed to are shown as a prefix following the format Fig# followed by a short description. All files are in .tsv format to ensure compatibility across different data analysis platforms.